refactor: useCouponModal hook 분리
